simple_switch_grpc: make read PEM reusable
This patch moves `read_pem_file()` and `read_pem_exception` under
`include` to enable reusability between targets. In particular, this
change would allow to reuse this code with the psa_switch target.
