+++ targets/simple_switch_grpc/main.cpp
@@ -21,6 +21,7 @@
 
 #include <bm/bm_sim/options_parse.h>
 #include <bm/bm_sim/target_parser.h>
+#include <bm/bm_grpc/pem.h>
 
 #include <exception>
 #include <fstream>
@@ -31,39 +32,6 @@
 #include <string>
 
 #include "switch_runner.h"
-
-namespace {
-
-class read_pem_exception : public std::exception {
- public:
-  read_pem_exception(const std::string &filename, const std::string &error)
-      : filename(filename), error(error) { }
-
-  std::string msg() const {
-    std::stringstream ss;
-    ss << "Error when reading pem file '" << filename << "': " << error << "\n";
-    return ss.str();
-  }
-
-  const char *what() const noexcept override {
-    return error.c_str();
-  }
-
- private:
-  std::string filename;
-  std::string error;
-};
-
-std::string read_pem_file(const std::string &filename) {
-  std::ifstream fs(filename, std::ios::in);
-  if (!fs) {
-    throw read_pem_exception(filename, "file cannot be opened");
-  }
-  return std::string((std::istreambuf_iterator<char>(fs)),
-                     std::istreambuf_iterator<char>());
-}
-
-}  // namespace
 
 int
 main(int argc, char* argv[]) {
@@ -233,17 +201,17 @@ main(int argc, char* argv[]) {
   try {
     if (grpc_server_ssl) {
       if (grpc_server_cacert != "") {
-        ssl_options->pem_root_certs = read_pem_file(grpc_server_cacert);
+        ssl_options->pem_root_certs = bm::read_pem_file(grpc_server_cacert);
       }
       if (grpc_server_cert != "") {
-        ssl_options->pem_cert_chain = read_pem_file(grpc_server_cert);
+        ssl_options->pem_cert_chain = bm::read_pem_file(grpc_server_cert);
       }
       if (grpc_server_key != "") {
-        ssl_options->pem_private_key = read_pem_file(grpc_server_key);
+        ssl_options->pem_private_key = bm::read_pem_file(grpc_server_key);
       }
       ssl_options->with_client_auth = grpc_server_with_client_auth;
     }
-  } catch (const read_pem_exception &e) {
+  } catch (const bm::read_pem_exception &e) {
     std::cerr << e.msg();
     std::exit(1);
   }
