+++ tools/check_style.sh
@@ -17,6 +17,7 @@ run_cpplint $ROOT_DIR/src/bm_sim src
 run_cpplint $ROOT_DIR/src/bm_apps src
 run_cpplint $ROOT_DIR/include/bm/bm_sim include
 run_cpplint $ROOT_DIR/include/bm/bm_apps include
+run_cpplint $ROOT_DIR/include/bm/bm_grpc include
 run_cpplint $ROOT_DIR/targets/simple_switch targets
 run_cpplint $ROOT_DIR/targets/l2_switch targets
 run_cpplint $ROOT_DIR/targets/simple_router targets
